requirements.txt
numpy==2.3.4
PyYAML==6.0.3
torch==2.9.0
torchvision==0.24.0
